# docs/api/openapi.core.v1.yaml
openapi: 3.1.0
info:
  title: DELMOS Core First Slice API
  version: 0.1.0-draft
  description: >-
    Target contract for the first Core role and approval slice. Not implemented.
    Session establishment and external integrations are outside this fragment.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /projects/{project_id}/milestones/{milestone_key}/accept:
    post:
      operationId: acceptMilestone
      summary: Record an independent gate decision for a project milestone
      x-required-permission: milestone.accept
      description: >-
        Accepts the milestone against the applied plan revision and the exact
        approved revisions of all required deliverables. The approver cannot
        be the author of the plan or any required deliverable. Returns the
        same immutable decision on retry. Unsupported acceptance rules and
        waivers do not silently pass.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: milestone_key
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Accepted gate decision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GateDecision' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Milestone not found }
        '422': { description: Plan not applied, missing evidence, self-decision, unsupported rule or mandatory rule veto }
  /projects/{project_id}/work-products/{work_product_id}/submit:
    post:
      operationId: submitWorkProduct
      summary: Submit the latest revision for review (CORE-CONTRACT-001, ADR-009)
      x-required-permission: wp.submit
      description: >-
        Creates a ReviewRequest bound to the exact (revision_id, payload_hash)
        pair and moves the work product to in_review. Assignees are passed
        explicitly: a permission grants the ability to act, an assignment
        defines what it applies to (ADR-009 §1). The author of the revision may
        not be assigned as reviewer or approver. Committing a new revision
        supersedes the open request, because decisions are never carried over
        to a new hash (ADR-009 §3).
      parameters:
        - name: project_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: work_product_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignments]
              properties:
                assignments:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [user_id, assignment_role]
                    properties:
                      user_id: { type: string, format: uuid }
                      assignment_role:
                        type: string
                        enum: [reviewer, approver]
      responses:
        '201':
          description: Review request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewRequest'
        '400': { description: Malformed body or assignment role }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Work product not found }
        '409': { description: An open review request already exists }
        '422': { description: Not a draft, no revision, assignee is the author or mandatory rule veto }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: delmos_session
      description: Server-side HttpOnly session; its creation is outside this fragment.
  parameters:
...
    ProjectId:
      in: path
      name: project_id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Missing permission for a visible action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ReviewRequest:
      type: object
      required: [id, project_id, work_product_id, revision_id, status, requested_by, created_at]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        work_product_id: { type: string, format: uuid }
        revision_id:
          type: string
          format: uuid
          description: The exact revision under review. Never reassigned to a newer one.
        status:
          type: string
          enum: [open, approved, changes_requested, superseded]
        requested_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
...
    GateDecision:
      type: object
      required: [id, milestone_key, outcome, config_generation]
      properties:
        id: { type: string, format: uuid }
        milestone_key: { type: string }
        outcome: { type: string, enum: [passed] }
        config_generation: { type: integer, format: int64 }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }